+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validators need now to specify the range of inputs of each claim they submit on-chain
 - Refactor GraphQL API for simplified architecture
 - Bumped grpc-interfaces to 0.11.0
+- Changed inspect-server to use new server-manager interface
 
 ### Removed